refactor(colors): share color/WCAG helpers via @formbricks/types (fix Sonar duplication)
The contrast helpers were duplicated across apps/web/lib/utils/colors.ts and
packages/surveys/src/lib/color.ts, tripping the SonarCloud "duplication on new
code" gate. Move the shared color math into @formbricks/types/colors (a
dependency of both) and turn the two files into thin re-exports; drop the
duplicated helper tests in the survey renderer (covered by apps/web) and the
redundant exhaustive sweep. Pure functions, so the widget bundle tree-shakes
what it doesn't use (node-html-parser stays externalized).
