+++ apps/web/lib/utils/colors.test.ts
@@ -101,16 +101,13 @@ describe("Color utilities", () => {
       );
     });
 
-    test("always returns an AA-compliant text color for any surface", () => {
-      for (let r = 0; r <= 255; r += 51) {
-        for (let g = 0; g <= 255; g += 51) {
-          for (let b = 0; b <= 255; b += 51) {
-            const surface = "#" + [r, g, b].map((x) => x.toString(16).padStart(2, "0")).join("");
-            expect(getContrastRatio(getReadableTextColor(surface), surface)).toBeGreaterThanOrEqual(
-              AA_CONTRAST_RATIO
-            );
-          }
-        }
+    test("always returns an AA-compliant text color across representative surfaces", () => {
+      // Mix of poles, saturated hues, and the mid-tones where black/white cross over near AA.
+      const surfaces = ["#ffffff", "#000000", "#ff0000", "#00ff00", "#0000ff", "#7f7f7f", "#808080", "#f9a8c0", "#ffff00", "#1e40af"]; // prettier-ignore
+      for (const surface of surfaces) {
+        expect(getContrastRatio(getReadableTextColor(surface), surface)).toBeGreaterThanOrEqual(
+          AA_CONTRAST_RATIO
+        );
       }
     });
   });
